+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
   - Process type should be globally pre-configured before application startup (new configuration options for each type of process);
   - Notifier can be restarted via `EvilEvents::Application.restart_event_notifier` (it would be convenient to have an ability to reload/restart notifier object in development mode interactively).
 - New more convenient exception hierarchy: now all exceptions live under `EvilEvents` namespace.
+- Added `AbstractEvent.emit!(**event_attributes)` method to provide an ability to emit events via event class constant.
 
 ## [0.2.0] - 2017-11-19
 ### Added